+++ lib/tailwind_ui.dart
@@ -1,6 +1,7 @@
 library tailwind_ui;
 
 export 'src/application_ui/elements/avatars/avatar_group_bottom_to_top.dart';
+export 'src/application_ui/elements/avatars/avatar_group_top_to_bottom.dart';
 export 'src/application_ui/elements/avatars/circular_avatar.dart';
 export 'src/application_ui/elements/avatars/circular_avatar_with_notification.dart';
 export 'src/application_ui/elements/avatars/circular_avatar_with_placeholder_icon.dart';